services/inference/requirements.txt
boto3
fastapi
mlflow
pillow
pydantic
torch
torchvision
uvicorn
